# Ensure fattest.simpliciy jar updates