--- a/dev/fattest.simplicity/build.gradle
+++ b/dev/fattest.simplicity/build.gradle
@@ -12,6 +12,29 @@
  *******************************************************************************/
 import org.apache.tools.ant.filters.ReplaceTokens
 
+configurations {
+	projectDependencies { transitive = false }
+}
+
+dependencies {
+	projectDependencies project(':com.ibm.websphere.javaee.jaxb.2.2'),
+	                    project(':com.ibm.websphere.javaee.jsonp.1.0'),
+	                    project(':com.ibm.ws.common.encoder'),
+	                    project(':com.ibm.ws.componenttest'),
+	                    project(':com.ibm.ws.jaxb.tools'),
+	                    project(':com.ibm.ws.junit.extensions'),
+	                    project(':com.ibm.ws.logging.core'),
+	                    project(':com.ibm.ws.org.glassfish.json'),
+	                    project(':com.ibm.ws.org.slf4j.api'),
+	                    project(':com.ibm.ws.org.slf4j.simple'),
+	                    project(':io.openliberty.com.fasterxml.jackson'),
+	                    project(':io.openliberty.org.apache.commons.logging'),
+	                    project(':io.openliberty.org.bouncycastle.bcpkix-jdk18on'),
+	                    project(':io.openliberty.org.bouncycastle.bcprov-jdk18on'),
+	                    project(':io.openliberty.org.bouncycastle.bcutil-jdk18on'),
+	                    project(':io.openliberty.org.testcontainers')	               
+}
+
 task assembleBootstrap(type: Copy) {
     inputs.property("javaHome", System.properties.('java.home').replace('\\', '\\\\').replace(':', '\\:'))
     def javaHome = System.properties.('java.home').replace('\\', '\\\\').replace(':', '\\:')
@@ -20,73 +43,39 @@ task assembleBootstrap(type: Copy) {
     filter(ReplaceTokens, tokens: [JAVA_HOME: javaHome])
 }
 
-task assembleBinaryDependencies() {
-    dependsOn ':com.ibm.websphere.javaee.jaxb.2.2:jar'
-    dependsOn ':com.ibm.websphere.javaee.jsonp.1.0:jar'
-    dependsOn ':com.ibm.ws.common.encoder:jar'
-    dependsOn ':com.ibm.ws.componenttest:jar'
-    dependsOn ':com.ibm.ws.jaxb.tools:jar'
-    dependsOn ':com.ibm.ws.junit.extensions:jar'
-    dependsOn ':com.ibm.ws.logging.core:jar'
-    dependsOn ':com.ibm.ws.org.glassfish.json:jar'
-    dependsOn ':com.ibm.ws.org.slf4j.api:jar'
-    dependsOn ':com.ibm.ws.org.slf4j.simple:jar'
-    dependsOn ':io.openliberty.com.fasterxml.jackson:jar'
-    dependsOn ':io.openliberty.org.apache.commons.logging:jar'
-    dependsOn ':io.openliberty.org.bouncycastle.bcpkix-jdk18on:jar'
-    dependsOn ':io.openliberty.org.bouncycastle.bcprov-jdk18on:jar'
-    dependsOn ':io.openliberty.org.bouncycastle.bcutil-jdk18on:jar'
-    dependsOn ':io.openliberty.org.testcontainers:jar'
-    File buildDirLib = new File(buildDir, 'lib')
-    outputs.dir buildDirLib
-    doLast {
-        // Copy maven dependencies of this project into a directory that can be included by the junit
-        // task of every FAT bucket, effectively making all of fattest.simplicty's dependencies transitive
-        sourceSets.main.runtimeClasspath.each {
-            def path = it.getAbsolutePath()
-            if((path.contains('.m2') || path.contains('.ibmartifactory') || path.contains('.ibmdhe')) && 
-               !(new File(buildDirLib, it.getName()).exists())) {
-                copy {
-                    from path
-                    into buildDirLib
-                }
-            }
-        }
-        copy {
-            from new File(project(':com.ibm.websphere.javaee.jaxb.2.2').buildDir, 'com.ibm.websphere.javaee.jaxb.2.2.jar')
-            from new File(project(':com.ibm.websphere.javaee.jsonp.1.0').buildDir, 'com.ibm.websphere.javaee.jsonp.1.0.jar')
-            from new File(project(':com.ibm.ws.componenttest').buildDir, 'com.ibm.ws.componenttest.jar')
-            from new File(project(':com.ibm.ws.jaxb.tools').buildDir, 'com.ibm.ws.jaxb.tools.jar')
-            from new File(project(':com.ibm.ws.junit.extensions').buildDir, 'com.ibm.ws.junit.extensions.jar')
-            from new File(project(':com.ibm.ws.logging.core').buildDir, 'com.ibm.ws.logging.core.jar')
-            from new File(project(':com.ibm.ws.org.glassfish.json').buildDir, 'com.ibm.ws.org.glassfish.json.1.0.jar')
-            from new File(project(':com.ibm.ws.org.slf4j.api').buildDir, 'com.ibm.ws.org.slf4j.api.jar')
-            from new File(project(':com.ibm.ws.org.slf4j.simple').buildDir, 'com.ibm.ws.org.slf4j.simple.jar')
-            from new File(project(':io.openliberty.com.fasterxml.jackson').buildDir, 'io.openliberty.com.fasterxml.jackson.jar')
-            from new File(project(':io.openliberty.org.apache.commons.logging').buildDir, 'io.openliberty.org.apache.commons.logging.jar')
-            from new File(project(':io.openliberty.org.bouncycastle.bcpkix-jdk18on').buildDir, 'io.openliberty.org.bouncycastle.bcpkix-jdk18on.jar')
-            from new File(project(':io.openliberty.org.bouncycastle.bcprov-jdk18on').buildDir, 'io.openliberty.org.bouncycastle.bcprov-jdk18on.jar')
-            from new File(project(':io.openliberty.org.bouncycastle.bcutil-jdk18on').buildDir, 'io.openliberty.org.bouncycastle.bcutil-jdk18on.jar')
-            from new File(project(':io.openliberty.org.testcontainers').buildDir, 'io.openliberty.org.testcontainers.jar')
-            from new File(buildDir, 'fattest.simplicity.jar')
-            into buildDirLib
-        }
-        copy {
-            from cnf.file('mavenlibs')
-            into buildDirLib
-            include 'infra.buildtasks-core-7.0.2.jar'
-            include 'asm-all-5.2.jar'
-            include 'org.apache.aries.util-*.jar'
-            include 'osgi.core*.jar'
-            include 'jsch-*.jar'
-        }
-        copy {
-            from project(':com.ibm.ws.common.encoder').buildDir
-            into buildDirLib
-            include 'com.ibm.ws.common.encoder.jar'
-            rename 'com.ibm.ws.common.encoder.jar', 'fattest.encoder.jar'
-        }
-    }
+task assembleBinaryDependencies(type: Copy) {
+	dependsOn ':cnf:copyMavenLibs'
+	
+	into(new File(buildDir, 'lib'))
+	
+	//fattest.simplicity.jar
+	from(jar) {
+	    exclude 'buildfiles'
+	}
+	
+	//Project dependencies
+	from(configurations.projectDependencies) {
+		rename 'com.ibm.ws.common.encoder.jar', 'fattest.encoder.jar'
+	}
+	
+	//Maven dependencies
+	from(cnf.file('mavenlibs')) {
+        include 'infra.buildtasks-core-7.0.2.jar'
+        include 'asm-all-5.2.jar'
+        include 'org.apache.aries.util-*.jar'
+        include 'osgi.core*.jar'
+        include 'jsch-*.jar'
+	}
+	
+	//Transitive dependencies (not from projects)
+	from(sourceSets.main.runtimeClasspath.findAll {
+		it.getAbsolutePath().contains('.m2') ||
+		it.getAbsolutePath().contains('.ibmartifactory') ||
+		it.getAbsolutePath().contains('.ibmdhe')
+	})
+	
+	// Debug task
+	// eachFile { println it.sourcePath }
 }
 
 configurations {