- [x] I have considered the risk of behavior change or other zero migration impact (https://github.com/OpenLiberty/open-liberty/wiki/Behavior-Changes).
- [x] If this PR fixes an Issue, the description includes "Fixes #FILLMEIN" or "Resolves #FILLMEIN" (verify `release bug` label if applicable: https://github.com/OpenLiberty/open-liberty/wiki/Open-Liberty-Conventions).
- [x] If this PR resolves an external Known Issue (including APARS), the description includes "Fixes #FILLMEIN" or "Resolves #FILLMEIN".

- This should fix the issue where switching between branches and running a test bucket doesn't copy the `fattest.simplicy:jar` into `fattest.simplicity/build/libs/lib`
  - This was achieved by using `from jar` in a copy task instead of `from new File(buildDir, 'fattest.simplicity.jar')`
  - Gradle checks what it needs to copy during the configuration stage.  If it sees that fattest.simpliciy.jar exists in both the to and from location it will skip the copy step.
  - When using `from jar` during the configuration stage gradle will check if `:jar` is going to be executed and then knows it needs to execute the copy as well.
- Made sure that this change kept the contents of `fattest.simplicity/build/libs/lib` the same after :clean :build and :clean :assemble
- Also verified performance of change is the same or better

## Before change

### Clean build

BUILD SUCCESSFUL in 34s
201 actionable tasks: 26 executed, 175 up-to-date

BUILD SUCCESSFUL in 32s
201 actionable tasks: 26 executed, 175 up-to-

BUILD SUCCESSFUL in 31s
201 actionable tasks: 26 executed, 175 up-to-date

### Build - no updates

BUILD SUCCESSFUL in 7s
201 actionable tasks: 18 executed, 183 up-to-date

BUILD SUCCESSFUL in 6s
201 actionable tasks: 18 executed, 183 up-to-date

BUILD SUCCESSFUL in 6s
201 actionable tasks: 18 executed, 183 up-to-date

### Build - recompile fattest.simplicity

BUILD SUCCESSFUL in 35s
201 actionable tasks: 21 executed, 180 up-to-date

BUILD SUCCESSFUL in 34s
201 actionable tasks: 21 executed, 180 up-to-date

BUILD SUCCESSFUL in 30s
201 actionable tasks: 21 executed, 180 up-to-date

## After change

### Clean build

BUILD SUCCESSFUL in 32s
204 actionable tasks: 29 executed, 175 up-to-date

BUILD SUCCESSFUL in 32s
204 actionable tasks: 29 executed, 175 up-to-date

BUILD SUCCESSFUL in 32s
204 actionable tasks: 29 executed, 175 up-to-date

### Build - no updates

BUILD SUCCESSFUL in 5s
204 actionable tasks: 18 executed, 186 up-to-date

BUILD SUCCESSFUL in 6s
204 actionable tasks: 18 executed, 186 up-to-date

BUILD SUCCESSFUL in 6s
204 actionable tasks: 18 executed, 186 up-to-date

### Build - recompile fattest.simplicity

BUILD SUCCESSFUL in 30s
204 actionable tasks: 22 executed, 182 up-to-date

BUILD SUCCESSFUL in 33s
204 actionable tasks: 22 executed, 182 up-to-date

BUILD SUCCESSFUL in 31s
204 actionable tasks: 22 executed, 182 up-to-date