@KyleAure 2024-07-20
Build failed with two known defects, and two intermittent network issues.